Add files via upload
Random Password generated of user_defined length, also included is a function to attempt to break it through brute force. If you want to test how good Google randomly generated passwords are then run this program and wait a very long time because the time complexity is O(k^n) with k representing the number of possible characters and n being the password length.